requirements.txt
pyTelegramBotAPI==4.4.1
python-dotenv==0.20.0
loguru==0.6.0
peewee==3.15.1
Pillow==9.2.0
requests==2.27.1
